matrix viewer fails on mesh with < ~10 elms
When running the small box mesh with '-snes view' petsc reports the
following error:

PETSC ERROR: Must call MatAssemblyBegin/End() before viewing matrix

The guess is that matrices for small meshes are not assembled.  Increasing the
element count results in a call to matrix assembly and the run
completing successfully with the option '-snes view'.